Fix docstring of alpha in dict_learning function

+++ sklearn/decomposition/_dict_learning.py
@@ -552,7 +552,7 @@ def dict_learning(
     n_components : int
         Number of dictionary atoms to extract.
 
-    alpha : int
+    alpha : float
         Sparsity controlling parameter.
 
     max_iter : int, default=100
